% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{outbreaks}
\alias{outbreaks}
\title{Infectious disease outbreaks in the USA, data from the CDC}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1642 rows and 12 columns.}
\source{
https://www.kaggle.com/cdc/foodborne-diseases
}
\usage{
outbreaks
}
\description{
This dataset provides data on foodborne disease outbreaks reported to CDC from 1998 
through 2015. Data fields include year, state (outbreaks occurring in more than one state are listed
as "multistate"), location where the food was prepared, reported food vehicle and contaminated 
ingredient, etiology (the pathogen, toxin, or chemical that caused the illnesses), status (whether 
the etiology was confirmed or suspected), total illnesses, hospitalizations, and fatalities. 
In many outbreak investigations, a specific food vehicle is not identified; for these outbreaks, 
the food vehicle variable is blank.
}
\keyword{datasets}
